Split logic even more
